requirements.txt
pandas
numpy 
openai
python-dotenv
chromadb
flask